requirements.txt
fire
pandas
pillow
playwright
python-dotenv
requests